+++ GameModule/src/Deck.h
@@ -17,9 +17,9 @@ public:
 	*/
 	Deck()
 	{
-		for (int i = 1; i <= 4; ++i)
+		for (size_t i = 1; i <= 4; ++i)
 		{
-			for (int j = 2; j <= 14; ++j)
+			for (size_t j = 2; j <= 14; ++j)
 			{
 				this->cards.push_back(new Card(i, j));
 			}
